requirements.txt
numpy==1.23.5
opencv-python==4.8.0.76
tensorflow==2.11.0
scikit-learn==1.1.3
matplotlib==3.6.2
Pillow>=9.0.0